A second production bug of the NOW() class: ::int in the free-tier cap
routes/projects.ts counted a founder's projects with `SELECT COUNT(*)::int
AS n`. `::` is Postgres cast syntax; D1 is SQLite and answers
`unrecognized token: ":"`.

The cap therefore THREW for every free founder who already had a
`founders` row — anyone creating a second project, or a first one after
onboarding had registered them. The core founder action returned a 500
instead of either the project or the clean 402 the cap exists to raise.
Verified against real SQLite: the current query throws, the fixed one
returns n=1 and correctly trips `count >= limit`.

Nothing could have caught it. tsc cannot see inside a SQL string, and no
test ran the query.

So the scanner that found it is now a guard. check-sqlite-dialect.mjs
fails the build on 21 constructs SQLite lacks — NOW(), GETDATE, DATE_ADD,
INTERVAL 'n', ISNULL, SELECT TOP, ILIKE, SERIAL, nextval, STRING_AGG,
TO_CHAR, EXTRACT(x FROM), ::casts and more — each with the dialect it came
from and the SQLite equivalent.

It scans SQL STRINGS, NOT FILES. A first pass grepped whole files and
reported 288 NOW() hits: Date.now(), performance.now(). `AGE(` matched
`MESSAGE(`, `TOP ` matched `TOPIC`, `SERIAL` matched `SERIALIZE`. Matching
on the name rather than the material is the failure this repo has
corrected repeatedly; the scanner extracts prepare(`…`), sql`…` and
exec(…) bodies and searches only inside those. 4083 SQL strings, and after
this fix, zero foreign dialect.

Injected and caught: capital.ts's NOW() restored, projects.ts's ::int
restored, and a fresh `DATE_SUB(NOW(), INTERVAL 7 DAY)` that trips three
patterns at once.

Six behaviour tests over a real database pin the cap itself, because the
next version of this bug will be a query that parses fine and counts the
wrong rows: all statuses count (a cap ignoring archived projects lets a
founder park one and keep creating), another founder's projects do not,
and the limit is read from the constant rather than hardcoded to 1.

One of those tests failed its own first draft — `doesNotMatch(src,
/COUNT\(\*\)::/)` against raw source, tripping on the comment beside the
fix, which quotes the broken syntax to explain it. The same prose-matching
mistake this branch spent a pass removing, made again while removing it.
It failed LOUDLY rather than passing quietly: that is the difference
between doesNotMatch and match on unstripped source, and it is why the
barrel-export guard in the previous commit could wave through four deleted
exports while this one could not. It now scans the SQL strings.

+++ cloudflare-worker/src/routes/projects.ts
@@ -374,7 +374,13 @@ async function createProjectHandler(c: any) {
     }
     if (founderRow.length > 0) {
       const fid = founderRow[0].id;
-      const existing = await sql`SELECT COUNT(*)::int AS n FROM projects WHERE founder_id = ${fid}`;
+      // `COUNT(*)`, not `COUNT(*)::int`. `::` is Postgres cast syntax and D1 is
+      // SQLite, which rejects it outright with `unrecognized token: ":"`. This
+      // threw for every free founder who already had a `founders` row —
+      // anyone creating a second project, or a first one after onboarding had
+      // registered them — so the core founder action returned a 500 instead of
+      // either the project or the clean 402 the cap is supposed to raise.
+      const existing = await sql`SELECT COUNT(*) AS n FROM projects WHERE founder_id = ${fid}`;
       const count = Number(existing?.[0]?.n ?? 0);
       if (count >= FREE_TIER_LIMITS.projects) {
         await sql.end();
